Move all exported defaults to a single source file for easier stubbing on platforms other than linux.
